Round 1 of remove Repa refactor
