+++ HSVO.cabal
@@ -15,8 +15,10 @@ cabal-version:       >=1.10
 
 library
   hs-source-dirs:      src
-  exposed-modules:     Numeric.Algorithms.HSVO.SVM
-                       , Numeric.Algorithms.HSVO.Detail
+  exposed-modules:
+                       Numeric.Algorithms.HSVO.SVM  
+                     , Numeric.Algorithms.HSVO.Detail
+                     , Numeric.Algorithms.HSVO.Types
   build-depends:       base >= 4.7 && < 5,
                        repa >= 3.4.0,
                        vector >= 0.11,
